Nest User component inside App
Again, I've moved the User component higher in the file so that it's
scoped correctly, and nested it within the App at the appropriate place.

Nothing here that we haven't already seen before, but I accidentally
commited some debugging code which might be a little interesting:

```jsx
 <input type="text" id="username" className="form-control" defaultValue="rohanorton"/>
```

In this I'm setting the value of the input to be pre-populated with my
github username, that means that I was able to test whether the code was
still working without having to enter the username again. It's a very
simple trick to save a tiny bit of effort, but having things like that
in your repertoire can make writing code just that little bit easier :)