Change User to UserForm and make it a form
I was finding the name User a bit confusing here, so I changed it to
UserForm as that seemed more accurate.

I also thought it would be nice to be able to hit return and get
results, so I changed it to use a form element.

On of the gotchas of using forms in javascript is that they tend to want
to make some kind of HTTP request which results in a page refresh or
navigation. As a result I've had to use the `preventDefault` method on
the event.

Hopefully the rest of the changes are pretty self-evident