Create onSubmit prop for UserForm
Previously we've seen that you can instantiate react components with
properties, and these can even include functions. This little refactor
makes use of this to create an onSubmit function for the UserForm which
simply takes a url. I have then moved the functionality that was
previously happening inside UserForm to App. Moving this functionality
to the top level component will come in handy later.