Track user state in App component
Introducing a new concept here: state management.

In react we can have components keep track of some bit of data as state.
Everytime the state is updated, react will rerender the component using
the new state.

In this change we're not actually using the state, just ensuring it is
set.

We can see that it starts off as null in getInitialState, and upon
completion of getJson we set it to the response data.

One of the odd things that you might have noticed is that we've had to
set a variable called `that` to be equal to `this`, this is because the
anonymous function we're passing as a callback will have a different
`this` context. It's pretty confusing stuff if you've not come across it
before.

Some useful `this` resources:
- https://rainsoft.io/gentle-explanation-of-this-in-javascript/
- http://javascriptissexy.com/understand-javascripts-this-with-clarity-and-master-it/