Use App state to render UserDetails
Since we have the user stored in state we are now easily able to nest
the UserDetails component within the App component, no longer having to
use the ReactDOM.render method!

This means that the application has taken on a much more declarative
structure as opposed to an imperative structure.

One of the benefits of this is that we could now do something truly
ridiculous like...

```jsx
ReactDOM.render(
    <div>
      <App />
      <App />
      <App />
    <div>,

    document.getElementById('app')
);
```

... creating three independent version of our app on the same page!

Well, almost. If you actually try to do this you'll notice that each of
these apps will render the same user details as the first one... We'll
fix this issue next