Use ref instead of getElementById
So, in the previous commit I pointed out a bug: if you put three
instances of the App on the page they all seemed to fetch the data
relating to the content of the first input.

This bug was caused by using getElementById to get the input element.
Because we were creating three of the App components, it was creating
three different elements with the ID of 'username'. Since the DOM
expects IDs to be unique, asking for the element with a particular ID
was causing the dom to simply return the first element it found.

To fix this we have changed from using an ID to using a react ref
instead. This means that the element is scoped to the object in
particular as opposed to being scoped to all the elements in the DOM.

This link explains quite a bit about refs, though note that they're
using a newer syntax than I chose to use in this commit:
- https://facebook.github.io/react/docs/refs-and-the-dom.html